CAClient: fully migrate to pyepics
Instead of caproto's pyepics compatible lib, use pyepics directly.
This is motivated by pyepics project repo stronger activity and
bigger community. There should be minor to no drawbacks to this
application.

Although some renamings happened, since those are a drop-in replacements
and structure is unchanged, this is considered a non-breaking change.